+++ angular.json
@@ -42,7 +42,8 @@
             ],
             "scripts": [],
             "server": "src/main.server.ts",
-            "prerender": true
+            "prerender": true,
+            "namedChunks": true
           },
           "configurations": {
             "production": {
