+++ src/app/dialogs/export-image-dialog/export-image-dialog.component.scss
