+++ src/app/dialogs/open-dialog/open-dialog.component.scss
