+++ src/app/dialogs/settings-dialog/settings-dialog.component.scss
