+++ src/app/dialogs/zip-download-model-dialog/zip-download-model-dialog.component.scss
